+++ Utilities/NrrdIO/reorder.c
@@ -322,7 +322,7 @@ nrrdShuffle(Nrrd *nout, const Nrrd *nin, unsigned int axis,
             me, axis, nin->dim-1);
     biffAdd(NRRD, err); return 1;
   }
-  len = nin->axis[axis].size;
+  len = (unsigned int)(nin->axis[axis].size);
   for (ai=0; ai<len; ai++) {
     if (!( perm[ai] < len )) {
       sprintf(err, "%s: perm[%d] (" _AIR_SIZE_T_CNV
@@ -381,7 +381,7 @@ nrrdShuffle(Nrrd *nout, const Nrrd *nin, unsigned int axis,
   memset(cOut, 0, (NRRD_DIM_MAX+1)*sizeof(int));
   for (idxOut=0; idxOut<numLines; idxOut++) {
     memcpy(cIn, cOut, ldim*sizeof(int));
-    cIn[0] = perm[cOut[0]];
+    cIn[0] = (unsigned int)(perm[cOut[0]]);
     NRRD_INDEX_GEN(idxIn, cIn, lsize, ldim);
     NRRD_INDEX_GEN(idxOut, cOut, lsize, ldim);
     memcpy(dataOut + idxOut*lineSize, dataIn + idxIn*lineSize, lineSize);
